Convert to Streamlined CoffeeScript.
Convert the sever to streamline.js based CoffeeScript. Most of this code is
already outgoing, since the matching of URLs to tests to run against their pages
has been moved into the browser extensions. The `model_.coffee` file will be
purged in an upcomming commit.

Also, the README was converted to Markdown. Needs more attention. It looks
funkyn on GitHub.