Remove invocation of parseInt.
A call to `parseInt` is causing problems in Internet Explorer because we're
running the test fetching JavaScript in a context outside of the web browser. We
run it in our own instance of the Active Scripting JScript engine.

The block of code that contains the JavaScript is outgoing in any case. We will
provide the expected count of test in the Verity JavaScript, not the metadata
comments.

It is simple enough to reference the `window.parseInt` of the browser, but we're
going to simply remove this.