Remove parseInt from common JavaScript.
Because we run the common JavaScript in an instnace of the JScript Active
Scripting engine of our own creation, `parseInt` is not available in the global
scope. We would have to alias an implementation of `parseInt`, probably from the
Window we hooked, through its `IWebBrowser2` interface.

We don't actually need the block of code that called `parseInt`, so we're
removing it. If we need `parseInt` or any of the other built-in functions, we'll
revisit this.