Removing Pool.h and Pool.c.
Initially the construction of the injected script was done in C, prior to
investing the time to create COM helper functions and macros, prior to the
decision to reuse the JavaScript used by all the other extensions.

Pool.c implemented a memory pool, a linked list of resources allocated during a
page visit, that could be deallocated all at once, once we'd run our test, or
else determined that a test was not necessary.

Since then, we've grown accustomed to COM (once again). We'll count on COM
reference counting to release our resources. Allocated resources will be owned
by COM object, and their memory released in the `Release` method of the COM
object.